Apply new scheme to flp2epn
Cmake changed.
structure changed.
code reformated.
imports optimized.
