+++ Examples/flp2epn/include/O2FLPex.h
@@ -15,18 +15,25 @@
 class O2FLPex : public FairMQDevice
 {
   public:
-    enum {
-      InputFile = FairMQDevice::Last,
-      EventSize,
-      Last
+    enum
+    {
+        InputFile = FairMQDevice::Last,
+        EventSize,
+        Last
     };
+
     O2FLPex();
+
     virtual ~O2FLPex();
+
     void Log(int intervalInMs);
 
-    virtual void SetProperty(const int key, const std::string& value);
-    virtual std::string GetProperty(const int key, const std::string& default_ = "");
+    virtual void SetProperty(const int key, const std::string &value);
+
+    virtual std::string GetProperty(const int key, const std::string &default_ = "");
+
     virtual void SetProperty(const int key, const int value);
+
     virtual int GetProperty(const int key, const int default_ = 0);
 
   protected:
