+++ Examples/flp2epn/include/O2Proxy.h
@@ -10,11 +10,11 @@
 
 #include "FairMQDevice.h"
 
-
-class O2Proxy: public FairMQDevice
+class O2Proxy : public FairMQDevice
 {
   public:
     O2Proxy();
+
     virtual ~O2Proxy();
 
   protected:
